@@ -54,47 +60,51 @@
     "pg-error-enum": "~0.5.0",
     "reflect-metadata": "~0.1.13",
     "rimraf": "~3.0.2",
-    "rxjs": "~6.6.7",
-    "st-utils": "~1.6.1",
+    "rxjs": "~7.1.0",
+    "st-utils": "~1.8.0",
     "swagger-ui-express": "~4.1.6",
-    "typeorm": "~0.2.32",
-    "typeorm-transactional-cls-hooked": "~0.1.20",
+    "typeorm": "~0.2.34",
+    "typeorm-transactional-cls-hooked": "~0.1.21",
     "url-metadata": "~2.5.0",
     "uuid": "~8.3.2"
   },
   "devDependencies": {
     "@nestjs/cli": "~7.6.0",
     "@nestjs/schematics": "~7.3.1",
-    "@nestjs/testing": "~7.6.15",
+    "@nestjs/testing": "~7.6.17",
+    "@types/adm-zip": "~0.4.34",
     "@types/bcryptjs": "~2.4.2",
     "@types/compression": "~1.7.0",
     "@types/config": "~0.0.38",
-    "@types/express": "~4.17.11",
-    "@types/express-rate-limit": "~5.1.1",
-    "@types/inline-css": "~0.0.33",
-    "@types/jest": "26.0.23",
+    "@types/express": "~4.17.12",
+    "@types/fs-extra": "~9.0.11",
+    "@types/inline-css": "~3.0.0",
+    "@types/jest": "~26.0.23",
     "@types/morgan": "~1.9.2",
-    "@types/node": "~14.14.41",
-    "@types/nodemailer": "~6.4.1",
+    "@types/node": "~14.14.45",
+    "@types/nodemailer": "~6.4.2",
     "@types/openid": "~2.0.1",
     "@types/passport-jwt": "~3.0.5",
     "@types/socket.io": "~2.1.13",
     "@types/supertest": "~2.0.11",
     "@types/uuid": "~8.3.0",
-    "@typescript-eslint/eslint-plugin": "~4.22.1",
-    "@typescript-eslint/parser": "~4.22.1",
-    "eslint": "~7.25.0",
+    "@typescript-eslint/eslint-plugin": "~4.26.0",
+    "@typescript-eslint/parser": "~4.26.0",
+    "adm-zip": "~0.5.5",
+    "eslint": "~7.28.0",
     "eslint-config-prettier": "~8.3.0",
-    "eslint-plugin-import": "~2.22.1",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "jest": "26.6.3",
-    "prettier": "~2.2.1",
+    "eslint-plugin-import": "~2.23.4",
+    "fs-extra": "~10.0.0",
+    "jest": "~26.6.3",
+    "ora": "~5.4.0",
+    "prettier": "~2.3.1",
     "supertest": "~6.1.3",
-    "ts-jest": "26.5.5",
+    "ts-jest": "~26.5.6",
     "ts-loader": "~8.1.0",
     "ts-node": "~9.1.1",
     "tsconfig-paths": "~3.9.0",
-    "typescript": "~4.2.4"
+    "typescript": "~4.3.2",
+    "yargs": "~17.0.1"
   },
   "jest": {
     "moduleFileExtensions": [
