@@ -1,12 +1,13 @@
 import { WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 import { Server } from 'socket.io';
 import { AuthSteamLoginSocketViewModel } from './auth.view-model';
+import { environment } from '../environment/environment';
 
 export enum AuthGatewayEvents {
   loginSteam = 'login-steam',
 }
 
-@WebSocketGateway({ namespace: `auth` })
+@WebSocketGateway({ path: environment.websocketPath, transports: environment.websocketTransports, namespace: '/auth' })
 export class AuthGateway {
   @WebSocketServer() server!: Server;
 
