@@ -1,6 +1,7 @@
 import { NamingStategy } from './naming.strategy';
 import { environment } from './environment';
 import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+import { resolve } from 'path';
 
 export const DB_TYPEORM_CONFIG: TypeOrmModuleOptions = {
   host: environment.get('DB_HOST'),
@@ -10,8 +11,10 @@ export const DB_TYPEORM_CONFIG: TypeOrmModuleOptions = {
   database: environment.get('DB_DATABASE'),
   synchronize: environment.get('DB_SYNCHRONIZE'),
   type: 'postgres',
-  entities: [__dirname + '/../**/*.entity{.ts,.js}'],
+  entities: [...environment.entitiesPaths],
   logging: !environment.production ? 'all' : false,
   namingStrategy: new NamingStategy(),
   dropSchema: false,
+  migrations: [resolve(process.cwd() + '/migration/*.js')],
+  cli: { migrationsDir: 'migration' },
 };
