@@ -8,10 +8,11 @@ export class MailService {
   constructor(private mailerService: MailerService) {}
 
   async sendMailInfo(options: ISendMailOptions, mailInfoTemplate: MailInfoTemplate): Promise<void> {
+    options = { ...options };
+    options.from ??= environment.mail;
     await this.mailerService.sendMail({
       ...options,
-      from: environment.get('MAIL'),
-      template: 'info',
+      template: './info.hbs',
       context: {
         title: mailInfoTemplate.title,
         version: environment.appVersion,
