@@ -4,6 +4,7 @@ import { PlatformGameMiniGame } from '../platform-game-mini-game/platform-game-m
 import { Mode } from '../../mode/mode.entity';
 import { PlatformGameMiniGameModeStage } from '../platform-game-mini-game-mode-stage/platform-game-mini-game-mode-stage.entity';
 import { Property } from '../../mapper/property.decorator';
+import { PlatformGameMiniGameModeCharacterCostume } from '../platform-game-mini-game-mode-character-costume/platform-game-mini-game-mode-character-costume.entity';
 
 @Unique(['idPlatformGameMiniGame', 'idMode'])
 @Entity()
@@ -13,7 +14,7 @@ export class PlatformGameMiniGameMode extends BaseEntity {
   idPlatformGameMiniGame!: number;
 
   @Property(() => PlatformGameMiniGame)
-  @ManyToOne(() => PlatformGameMiniGame)
+  @ManyToOne(() => PlatformGameMiniGame, platformGameMiniGame => platformGameMiniGame.platformGameMiniGameModes)
   @JoinColumn()
   platformGameMiniGame!: PlatformGameMiniGame;
 
@@ -32,4 +33,11 @@ export class PlatformGameMiniGameMode extends BaseEntity {
     platformGameMiniGameModeStage => platformGameMiniGameModeStage.platformGameMiniGameMode
   )
   platformGameMiniGameModeStages!: PlatformGameMiniGameModeStage[];
+
+  @Property(() => PlatformGameMiniGameModeCharacterCostume)
+  @OneToMany(
+    () => PlatformGameMiniGameModeCharacterCostume,
+    platformGameMiniGameModeCharacterCostume => platformGameMiniGameModeCharacterCostume.platformGameMiniGameMode
+  )
+  platformGameMiniGameModeCharacterCostumes!: PlatformGameMiniGameModeCharacterCostume[];
 }
