@@ -5,11 +5,12 @@ import { PlayerAddDto, PlayerUpdateDto } from './player.dto';
 import { SteamService } from '../steam/steam.service';
 import { RegionService } from '../region/region.service';
 import { NotOrNull } from '../util/find-operator';
-import { PlayerViewModel, PlayerWithRegionViewModel } from './player.view-model';
+import { PlayerViewModel, PlayerWithRegionSteamProfileViewModel } from './player.view-model';
 import { MapperService } from '../mapper/mapper.service';
 import { Transactional } from 'typeorm-transactional-cls-hooked';
 import { ILike } from 'typeorm';
 import { Pagination } from 'nestjs-typeorm-paginate';
+import { isAfter, subDays } from 'date-fns';
 
 @Injectable()
 export class PlayerService {
@@ -61,7 +62,7 @@ export class PlayerService {
   }
 
   async findById(idPlayer: number): Promise<Player> {
-    const player = await this.playerRepository.findOne(idPlayer, { relations: ['region'] });
+    const player = await this.playerRepository.findOne(idPlayer, { relations: ['region', 'steamProfile'] });
     if (!player) {
       throw new NotFoundException('Player not found');
     }
@@ -72,18 +73,26 @@ export class PlayerService {
     return this.playerRepository.findOneOrFail(idPlayer, { relations: ['user'] });
   }
 
-  async findByIdMapped(idPlayer: number): Promise<PlayerWithRegionViewModel> {
+  async findByIdMapped(idPlayer: number): Promise<PlayerWithRegionSteamProfileViewModel> {
     const player = await this.findById(idPlayer);
-    return this.mapperService.map(Player, PlayerWithRegionViewModel, player);
+    return this.mapperService.map(Player, PlayerWithRegionSteamProfileViewModel, player);
   }
 
-  async update(idPlayer: number, dto: PlayerUpdateDto): Promise<PlayerWithRegionViewModel> {
+  async update(idPlayer: number, dto: PlayerUpdateDto): Promise<PlayerWithRegionSteamProfileViewModel> {
     const player = await this.playerRepository.findOneOrFail(idPlayer);
     await this.playerRepository.update(idPlayer, dto);
     if (dto.idRegion && player.idRegion !== dto.idRegion) {
       player.region = await this.regionService.findById(dto.idRegion);
     }
-    return this.mapperService.map(Player, PlayerWithRegionViewModel, new Player().extendDto({ ...player, ...dto }));
+    return this.mapperService.map(
+      Player,
+      PlayerWithRegionSteamProfileViewModel,
+      new Player().extendDto({ ...player, ...dto })
+    );
+  }
+
+  async updateIdUser(idPlayer: number, idUser: number): Promise<void> {
+    await this.playerRepository.update(idPlayer, { idUser });
   }
 
   async delete(idPlayer: number): Promise<void> {
@@ -98,7 +107,7 @@ export class PlayerService {
     return this.steamService.openIdUrl(player);
   }
 
-  async unlinkSteamProfile(idPlayer: number): Promise<PlayerWithRegionViewModel> {
+  async unlinkSteamProfile(idPlayer: number): Promise<PlayerWithRegionSteamProfileViewModel> {
     return this.update(idPlayer, { idSteamProfile: undefined });
   }
 
@@ -108,15 +117,6 @@ export class PlayerService {
 
   async findIdByIdUser(idUser: number): Promise<number> {
     return this.playerRepository.findOneOrFail({ select: ['id'], where: { idUser } }).then(player => player.id);
-  }
-
-  // TODO REMOVE
-  async findRandom(idsNot: number[]): Promise<Player> {
-    const qb = this.playerRepository.createQueryBuilder('p').orderBy('random()');
-    if (idsNot.length) {
-      qb.andWhere('p.id not in (:...ids)', { ids: idsNot });
-    }
-    return qb.getOneOrFail();
   }
 
   async findBySearch(
@@ -135,5 +135,25 @@ export class PlayerService {
 
   async personaNameExists(personaName: string): Promise<boolean> {
     return this.playerRepository.exists({ personaName });
+  }
+
+  async personaNameExistsWithoutUser(personaName: string): Promise<boolean> {
+    return this.playerRepository.exists({ personaName, noUser: true });
+  }
+
+  async updatePersonaName(idPlayer: number, personaName: string): Promise<string> {
+    const player = await this.playerRepository.findOneOrFail(idPlayer);
+    if (player.personaName === personaName) {
+      throw new BadRequestException('New personaName is the same as the old personaname');
+    }
+    if (await this.personaNameExists(personaName)) {
+      throw new BadRequestException('PersonaName already taken');
+    }
+    if (player.lastUpdatedPersonaNameDate && isAfter(player.lastUpdatedPersonaNameDate, subDays(new Date(), 7))) {
+      throw new BadRequestException(`Too many updates in the last 7 days`);
+    }
+    const lastUpdatedPersonaNameDate = new Date();
+    await this.playerRepository.update(idPlayer, { personaName, lastUpdatedPersonaNameDate });
+    return lastUpdatedPersonaNameDate.toISOString();
   }
 }
