@@ -17,10 +17,10 @@ import { ApiOrderByAndDir } from '../shared/order-by/api-order-by';
 import { OrderByDirection } from 'st-utils';
 import { ScoreChangeRequestsPaginationViewModel } from './view-model/score-change-request.view-model';
 import { ScoreChangeRequest } from './score-change-request/score-change-request.entity';
-import { ScoreStatusEnum } from './score-status.enum';
 import { Pagination } from 'nestjs-typeorm-paginate';
 import { ApiPagination } from '../shared/decorator/api-pagination';
-import { addDays } from 'date-fns';
+import { ScoreGroupedByStatusViewModel } from './view-model/score-grouped-by-status.view-model';
+import { ScoreStatusEnum } from './score-status/score-status.enum';
 
 @ApiAuth()
 @ApiTags('Score')
@@ -31,65 +31,6 @@ export class ScoreController {
   @Post()
   async add(@Body() dto: ScoreAddDto, @AuthUser() user: User): Promise<ScoreViewModel> {
     return this.scoreService.add(dto, user);
-  }
-
-  @ApiQuery({ name: 'platform', required: false })
-  @ApiQuery({ name: 'game', required: false })
-  @ApiQuery({ name: 'miniGame', required: false })
-  @ApiQuery({ name: 'mode', required: false })
-  @ApiQuery({ name: 'approved', required: false })
-  @ApiQuery({ name: 'characterCostume', required: false })
-  @ApiQuery({ name: 'stage', required: false })
-  @Post('insert-random')
-  // TODO REMOVE
-  async insertRandom(
-    @AuthUser() user: User,
-    @Query('platform') platform?: string,
-    @Query('game') game?: string,
-    @Query('miniGame') miniGame?: string,
-    @Query('mode') mode?: string,
-    @Query('approved') approved?: boolean,
-    @Query('characterCostume') characterCostume?: string,
-    @Query('stage') stage?: string
-  ): Promise<void> {
-    await this.scoreService.insert({ miniGame, platform, mode, game, approved, characterCostume, stage }, user);
-  }
-
-  @ApiQuery({ name: 'platform', required: false })
-  @ApiQuery({ name: 'game', required: false })
-  @ApiQuery({ name: 'miniGame', required: false })
-  @ApiQuery({ name: 'mode', required: false })
-  @ApiQuery({ name: 'approved', required: false })
-  @ApiQuery({ name: 'characterCostume', required: false })
-  @ApiQuery({ name: 'stage', required: false })
-  @ApiQuery({ name: 'fromDate', required: false })
-  @Post('insert-many-random')
-  // TODO REMOVE
-  async insertManyRandom(
-    @AuthUser() user: User,
-    @Query('q') q: number,
-    @Query('platform') platform?: string,
-    @Query('game') game?: string,
-    @Query('miniGame') miniGame?: string,
-    @Query('mode') mode?: string,
-    @Query('approved') approved?: boolean,
-    @Query('characterCostume') characterCostume?: string,
-    @Query('stage') stage?: string,
-    @Query('fromDate', OptionalQueryPipe) fromDate?: Date
-  ): Promise<void> {
-    if (fromDate) {
-      fromDate = new Date(fromDate);
-    }
-    for (let i = 0; i < q; i++) {
-      await this.scoreService.insert(
-        { miniGame, platform, mode, game, approved, characterCostume, stage },
-        user,
-        fromDate
-      );
-      if (fromDate) {
-        fromDate = addDays(fromDate, 1);
-      }
-    }
   }
 
   @ApiQuery({ name: Params.limit, required: false })
@@ -211,6 +152,11 @@ export class ScoreController {
     return this.scoreService.findScoresWithChangeRequestsCount(user);
   }
 
+  @Get('player/rejected-and-pending')
+  async findRejectedAndPendingScoresByIdUser(@AuthUser() user: User): Promise<ScoreGroupedByStatusViewModel[]> {
+    return this.scoreService.findRejectedAndPendingScoresByIdUser(user.id);
+  }
+
   @ApiQuery({ name: Params.worldRecord, required: false })
   @ApiQuery({ name: Params.characterWorldRecord, required: false })
   @ApiQuery({ name: Params.combinationWorldRecord, required: false })
@@ -221,7 +167,7 @@ export class ScoreController {
   @ApiQuery({ name: Params.idModes, required: false, isArray: true, type: Number })
   @ApiQuery({ name: Params.idStages, required: false, isArray: true, type: Number })
   @ApiQuery({ name: Params.idCharacterCostumes, required: false, isArray: true, type: Number })
-  @ApiQuery({ name: Params.status, enum: ScoreStatusEnum })
+  @ApiQuery({ name: Params.idScoreStatus, enum: ScoreStatusEnum, required: false })
   @ApiPagination(ScoreViewModel)
   @Get('search')
   async searchScores(@Query() dto: ScoreSearchDto, @AuthUser() user: User): Promise<Pagination<ScoreViewModel>> {
