@@ -1,11 +1,12 @@
 import { WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 import { Server } from 'socket.io';
+import { environment } from '../environment/environment';
 
 export enum ScoreGatewayEvents {
   updateCountApprovals = 'updateCountApprovals',
 }
 
-@WebSocketGateway({ namespace: 'score', transports: ['websockets'] })
+@WebSocketGateway({ path: environment.websocketPath, transports: environment.websocketTransports, namespace: '/score' })
 export class ScoreGateway {
   @WebSocketServer() server!: Server;
 
