@@ -15,10 +15,5 @@ describe('AppController (e2e)', () => {
     await app.init();
   });
 
-  it('/ (GET)', () => {
-    return request(app.getHttpServer())
-      .get('/')
-      .expect(200)
-      .expect('Hello World!');
-  });
+  it('/ (GET)', () => request(app.getHttpServer()).get('/').expect(200).expect('Hello World!'));
 });
